requirements.txt
streamlit
torch
timm
joblib
Pillow
matplotlib